requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.1
Django==3.0.4
importlib-metadata==1.5.0
kombu==4.6.8
python-dotenv==0.12.0
pytz==2019.3
redis==3.4.1
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
